requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.4
opencv-python==4.10.0.84
ultralytics==8.3.0
numpy==1.26.4
PyYAML==6.0.2